Defers event handling to app handlers.
