+++ src/introductions/mod.rs
@@ -1,3 +1,4 @@
 pub mod commands;
+pub mod handlers;
 pub mod queries;
 pub mod voice;
